superResolution/rcan/mtk/python/requirements.txt
torch==1.13.1
numpy==1.24.3
onnx==1.20.1
onnx-simplifier==0.4.36
onnxruntime==1.23.2
tensorflow==2.13.0
Pillow
